requirements.txt
immunity-controller @ https://github.com/edge-servers/immunity-controller/tarball/master
influxdb~=5.3.1
django-nested-admin~=4.0.2
netaddr~=0.8.0
python-dateutil>=2.7.0,<3.0.0
immunity-utils[rest] @ https://github.com/edge-servers/immunity-utils/tarball/master
